* use iso8601 library to parse durations
