+++ spring-boot-admin-server-ui/package.json
@@ -30,6 +30,7 @@
     "d3-time": "^2.0.0",
     "event-source-polyfill": "^1.0.21",
     "file-saver": "^2.0.2",
+    "iso8601-duration": "^1.3.0",
     "lodash": "^4.17.20",
     "moment": "^2.29.1",
     "popper.js": "^1.16.1",
